Fix AdMob initialization issues and implement MultiDex support

+++ app/src/main/java/com/example/barkodm/ui/inventory/InventoryDetailFragment.kt
@@ -42,7 +42,7 @@ class InventoryDetailFragment : Fragment() {
         super.onViewCreated(view, savedInstanceState)
         
         // Fragment argümanlarından envanter ID'sini al
-        val inventoryId = args.inventoryId
+        val inventoryId = args.inventoryId.toLong()
         
         // ViewModel'e envanter ID'sini aktar
         viewModel.setInventoryId(inventoryId)
